+++ src/infinity/facets/TokenFacet.sol
@@ -2,7 +2,6 @@
  * SPDX-License-Identifier: MIT
  *
  */
-
 pragma solidity ^0.8.16;
 
 import "../AppStorage.sol";
@@ -16,24 +15,23 @@ import "../libraries/Token/LibMatic.sol";
  */
 contract TokenFacet {
     struct Balance {
-        uint internalBalance;
-        uint externalBalance;
-        uint totalBalance;
+        uint256 internalBalance;
+        uint256 externalBalance;
+        uint256 totalBalance;
     }
 
     using SafeERC20 for IERC20;
 
-    event InternalBalanceChanged(address indexed user, IERC20 indexed token, int delta);
+    event InternalBalanceChanged(address indexed user, IERC20 indexed token, int256 delta);
 
     /**
      * Transfer
      *
      */
-
     function transferToken(
         IERC20 token,
         address recipient,
-        uint amount,
+        uint256 amount,
         LibTransfer.From fromMode,
         LibTransfer.To toMode
     ) external payable {
@@ -44,13 +42,12 @@ contract TokenFacet {
      * Weth
      *
      */
-
-    function wrapMatic(uint amount, LibTransfer.To mode) external payable {
+    function wrapMatic(uint256 amount, LibTransfer.To mode) external payable {
         LibWmatic.wrap(amount, mode);
         LibMatic.refundMatic();
     }
 
-    function unwrapMatic(uint amount, LibTransfer.From mode) external payable {
+    function unwrapMatic(uint256 amount, LibTransfer.From mode) external payable {
         LibWmatic.unwrap(amount, mode);
     }
 
@@ -61,47 +58,47 @@ contract TokenFacet {
 
     // Internal
 
-    function getInternalBalance(address account, IERC20 token) public view returns (uint balance) {
+    function getInternalBalance(address account, IERC20 token) public view returns (uint256 balance) {
         balance = LibBalance.getInternalBalance(account, token);
     }
 
     function getInternalBalances(address account, IERC20[] memory tokens)
         external
         view
-        returns (uint[] memory balances)
+        returns (uint256[] memory balances)
     {
         balances = new uint256[](tokens.length);
-        for (uint i; i < tokens.length; ++i) {
+        for (uint256 i; i < tokens.length; ++i) {
             balances[i] = getInternalBalance(account, tokens[i]);
         }
     }
 
     // External
 
-    function getExternalBalance(address account, IERC20 token) public view returns (uint balance) {
+    function getExternalBalance(address account, IERC20 token) public view returns (uint256 balance) {
         balance = token.balanceOf(account);
     }
 
     function getExternalBalances(address account, IERC20[] memory tokens)
         external
         view
-        returns (uint[] memory balances)
+        returns (uint256[] memory balances)
     {
         balances = new uint256[](tokens.length);
-        for (uint i; i < tokens.length; ++i) {
+        for (uint256 i; i < tokens.length; ++i) {
             balances[i] = getExternalBalance(account, tokens[i]);
         }
     }
 
     // Total
 
-    function getBalance(address account, IERC20 token) public view returns (uint balance) {
+    function getBalance(address account, IERC20 token) public view returns (uint256 balance) {
         balance = LibBalance.getBalance(account, token);
     }
 
-    function getBalances(address account, IERC20[] memory tokens) external view returns (uint[] memory balances) {
+    function getBalances(address account, IERC20[] memory tokens) external view returns (uint256[] memory balances) {
         balances = new uint256[](tokens.length);
-        for (uint i; i < tokens.length; ++i) {
+        for (uint256 i; i < tokens.length; ++i) {
             balances[i] = getBalance(account, tokens[i]);
         }
     }
@@ -120,7 +117,7 @@ contract TokenFacet {
         returns (Balance[] memory balances)
     {
         balances = new Balance[](tokens.length);
-        for (uint i; i < tokens.length; ++i) {
+        for (uint256 i; i < tokens.length; ++i) {
             balances[i] = getAllBalance(account, tokens[i]);
         }
     }
