+++ src/infinity/interfaces/IC3.sol
@@ -3,15 +3,15 @@
 pragma solidity ^0.8.0;
 
 interface IC3Pool {
-    function freeRedeem(uint amount) external;
+    function freeRedeem(uint256 amount) external;
 
-    function taxedRedeem(address[] memory erc20Addresses, uint[] memory amount) external;
+    function taxedRedeem(address[] memory erc20Addresses, uint256[] memory amount) external;
 
     function getFreeRedeemAddresses() external view returns (address[] memory);
 
     function getERC20Tokens() external view returns (address[] memory);
 
-    function feeRedeem() external view returns (uint);
+    function feeRedeem() external view returns (uint256);
 }
 
 interface IC3ProjectToken {
