+++ src/infinity/interfaces/ICarbonChain.sol
@@ -3,5 +3,5 @@
 pragma solidity ^0.8.4;
 
 interface ICarbonChain {
-    function offsetCarbon(uint _carbonTon, string calldata _transactionInfo, string calldata _onBehalfOf) external;
+    function offsetCarbon(uint256 _carbonTon, string calldata _transactionInfo, string calldata _onBehalfOf) external;
 }
