+++ src/infinity/interfaces/IKlima.sol
@@ -3,21 +3,21 @@
 pragma solidity ^0.8.16;
 
 interface IStaking {
-    function unstake(uint _amount, bool _trigger) external;
+    function unstake(uint256 _amount, bool _trigger) external;
 }
 
 interface IStakingHelper {
-    function stake(uint _amount) external;
+    function stake(uint256 _amount) external;
 }
 
 interface IwsKLIMA {
-    function wrap(uint _amount) external returns (uint);
+    function wrap(uint256 _amount) external returns (uint256);
 
-    function unwrap(uint _amount) external returns (uint);
+    function unwrap(uint256 _amount) external returns (uint256);
 
-    function wKLIMATosKLIMA(uint _amount) external view returns (uint);
+    function wKLIMATosKLIMA(uint256 _amount) external view returns (uint256);
 
-    function sKLIMATowKLIMA(uint _amount) external view returns (uint);
+    function sKLIMATowKLIMA(uint256 _amount) external view returns (uint256);
 }
 
 interface IKlimaRetirementBond {
