+++ src/infinity/interfaces/IKlimaCarbonRetirements.sol
@@ -6,21 +6,21 @@ interface IKlimaCarbonRetirements {
     function carbonRetired(
         address _retiree,
         address _pool,
-        uint _amount,
+        uint256 _amount,
         string calldata _beneficiaryString,
         string calldata _retirementMessage
     ) external;
 
-    function getUnclaimedTotal(address _minter) external view returns (uint);
+    function getUnclaimedTotal(address _minter) external view returns (uint256);
 
-    function offsetClaimed(address _minter, uint _amount) external returns (bool);
+    function offsetClaimed(address _minter, uint256 _amount) external returns (bool);
 
-    function getRetirementIndexInfo(address _retiree, uint _index)
+    function getRetirementIndexInfo(address _retiree, uint256 _index)
         external
         view
-        returns (address, uint, string memory, string memory);
+        returns (address, uint256, string memory, string memory);
 
-    function getRetirementPoolInfo(address _retiree, address _pool) external view returns (uint);
+    function getRetirementPoolInfo(address _retiree, address _pool) external view returns (uint256);
 
-    function getRetirementTotals(address _retiree) external view returns (uint, uint, uint);
+    function getRetirementTotals(address _retiree) external view returns (uint256, uint256, uint256);
 }
