+++ src/protocol/staking/regular/StakingHelper.sol
@@ -1,7 +1,5 @@
-
 // SPDX-License-Identifier: AGPL-3.0-or-later
 pragma solidity 0.7.5;
-
 
 interface IERC20 {
     function decimals() external view returns (uint8);
@@ -76,26 +74,25 @@ interface IERC20 {
 }
 
 interface IStaking {
-    function stake( uint _amount, address _recipient ) external returns ( bool );
-    function claim( address _recipient ) external;
+    function stake(uint256 _amount, address _recipient) external returns (bool);
+    function claim(address _recipient) external;
 }
 
 contract StakingHelper {
-
     address public immutable staking;
     address public immutable KLIMA;
 
-    constructor ( address _staking, address _KLIMA ) {
-        require( _staking != address(0) );
+    constructor(address _staking, address _KLIMA) {
+        require(_staking != address(0));
         staking = _staking;
-        require( _KLIMA != address(0) );
+        require(_KLIMA != address(0));
         KLIMA = _KLIMA;
     }
 
-    function stake( uint _amount ) external {
-        IERC20( KLIMA ).transferFrom( msg.sender, address(this), _amount );
-        IERC20( KLIMA ).approve( staking, _amount );
-        IStaking( staking ).stake( _amount, msg.sender );
-        IStaking( staking ).claim( msg.sender );
+    function stake(uint256 _amount) external {
+        IERC20(KLIMA).transferFrom(msg.sender, address(this), _amount);
+        IERC20(KLIMA).approve(staking, _amount);
+        IStaking(staking).stake(_amount, msg.sender);
+        IStaking(staking).claim(msg.sender);
     }
 }
