+++ test/helpers/AssertionHelper.sol
@@ -1,11 +1,12 @@
 // SPDX-License-Identifier: UNLICENSED
 pragma solidity ^0.8.0;
 
-/******************************************************************************\
-* Authors: Cujo <rawr@cujowolf.dev>
-* Helper functions for the common assertions used in testing
-/******************************************************************************/
-
+/**
+ * \
+ * Authors: Cujo <rawr@cujowolf.dev>
+ * Helper functions for the common assertions used in testing
+ * /*****************************************************************************
+ */
 import "forge-std/Test.sol";
 import "oz/token/ERC20/IERC20.sol";
 
@@ -14,7 +15,7 @@ abstract contract AssertionHelper is Test {
         assertEq(IERC20(token).balanceOf(target), 0);
     }
 
-    function assertTokenBalance(address token, address target, uint amount) internal {
+    function assertTokenBalance(address token, address target, uint256 amount) internal {
         assertEq(IERC20(token).balanceOf(target), amount);
     }
 }
